+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Notable changes to this project are documented in this file.
 # Unreleased
 
 ## Added
-- new command autoinstall intended for install ide without user interaction. 
+- new command 'autoinstall' intended for install ide without user interaction. 
+- new command 'defaults' to save user defaults (hostname only is supported yet)
 
 ## Changed
 - IDE install on first start runs in quick mode